Memory leak fix of sqlite driver when close failed
If an ongoing query is not finalized before close function is called,
sqlite driver can't close the connection to sqlite. In this case, the
failure is only reported to the client via getLastError(). Even if the
client want to handle this case, there is no easy way to revoke the
close function. This commit finalizes all queries before
close is called.

Task-number: QTBUG-16967
Change-Id: I069f6073bd46f53780e13d269e44aa86366e0574
Reviewed-by: Andy Shaw <andy.shaw@digia.com>
Reviewed-by: Mark Brand <mabrand@mabrand.nl>