Hang on to the correct screen index in QWidget
Hang onto the screen index that this widget should be
placed on in case the platform cannot move the window onto the screen yet,
when window is not yet created and the plugin cannot access the topData
to update the screenIndex. When this variable is later referenced
after the window is created, it defaults to screenIndex 0,
so the window is never moved correctly.

Change-Id: Ica78fd90aa3dacc29f81b2f6bd6c8b45d2fb2961
Reviewed-by: Jørgen Lind <jorgen.lind@nokia.com>