Fix q_atomic_decrement for 64-bit Solaris
q_atomic_decrement and q_atomic_increment did not set return value
correctly. They expected the %eax return value register to be empty,
when it was not necessarily so.

This could occur when when compiler inlined e.g.,
~QByteArray -> QBasicAtomicInt::deref -> q_atomic_decrement
chain to application code. Data would not to be freed and leak memory.

Fixed so that %eax is now zero'd before setting return value.

Task-number: QTBUG-24470
Change-Id: I83495e3ddb938713863a7b00714215c331b9562b
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>