Crash in Symbian text codec creation without CleanupStack
The Symbian text codec is creating a CCnvCharacterSetConverter with its
NewL() function. This requires the CleanupStack, otherwise you get a
E32USER-CBase 69 panic. But sometimes apps have been using text streams
and other objects in static data construction or destruction which
use the text codec where there is no CleanupStack.

This change detects is a CleanupStack is present and temporarily creates
one if not.

Change-Id: Ia68fa60b4f0e3d675ee302208a625c6e9493a6f1
Reviewed-by: Shane Kearns <shane.kearns@accenture.com>