Disable the use of sed on native mingw
Partially reverts 28f294e1e98f2437ce0d9614e698df7c8dc52dc6

Commit 28f294e1e98f2437ce0d9614e698df7c8dc52dc6
introduced the use of 'sed' to do replacements for mingw in
pkgconfig files, similar to what is done for unix.

However, this causes corrupted pkgconfig files when 'sed' is not
available. While on Unix we can safely assume that 'sed' is always
available, it appears that when building on Windows with mingw
'sed' might not be available.

As it would be too complicated for the 4.8 branch to add support
for automatic 'sed' detection, this patch chooses to completely
disable the use of 'sed' for all mingw builds on Windows.

For cross compiled builds 'sed' is still used.

Change-Id: I8b6b3f52df6906cd087dd3eb417bdb5800e8eafc
Reviewed-by: Andy Shaw <andy.shaw@digia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>