Upgrading libpng: Adding clean copy of libpng 1.5.10
This commit removes the previous version of the bundled
libpng (1.5.4), as well as all local modifications to it.
It adds an unmodified copy of the official libpng
source distribution, except that various extraneous
files have been removed, as usual.

The patches required to build it in Qt will follow in
separate commit(s).

Change-Id: I0cfc782a1c1b7f06b92eb1b38992bd8a0e6de08d
Reviewed-by: Kim M. Kalland <kim.kalland@nokia.com>