configure: Hack to not use the pg_config of the host system which will add /usr/include
pg_config is a native binary so using that when cross compiling
will always fail.  The commented out fix would do, but for OE
our -platform and -xplatform is the same so we are actually not
really cross compiling. Just comment out the test, we are passing
the location to the postgres headers and if they are okay we
will pad.

Upstream-Status: Inappropriate [configuration]

Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>